consolidate filter_tables with query_across_tables
