+++ synapsebridgehelpers/__init__.py
@@ -1,4 +1,4 @@
-from .tableHelpers import *
+from .tableHelpers import query_across_tables, get_tables, find_tables_with_data
 from .findHealthCodes import *
 from .filterTablesByActivity import *
 from .getFileIds import *
@@ -6,5 +6,4 @@ from .delAllTables import *
 from .transferTables import *
 from .tableStats import *
 from .summaryTable import *
-from .filter_tables import filter_tables
 from .export_tables import export_tables, compare_schemas, synchronize_schemas
